feat: implementar endpoint para pegar usuário logado

+++ src/app/user/user.component.ts
@@ -1,4 +1,4 @@
-import { ChangeDetectionStrategy, Component } from '@angular/core';
+import { Component, Input } from '@angular/core';
 import { MatIconModule } from "@angular/material/icon"
 import { NgIf } from "@angular/common";
 import { UserService } from './user.service';
@@ -11,25 +11,33 @@ import { User } from './user.model';
   imports: [MatIconModule, NgIf],
   templateUrl: './user.component.html',
   styleUrl: './user.component.css',
-  changeDetection: ChangeDetectionStrategy.OnPush,
 })
 export class UserComponent {
 
-  user!: User;
-  loading = true;
-  error = '';
+  @Input() user!: User | null;
+  @Input() loading = true;
+  @Input() error = '';
   
   constructor(private userService: UserService){}
 
-  onEdit() {
-    this.userService.editUser(this.user.id, this.user).subscribe({
-      next: () => alert("Usuário atualizado com sucesso"),
-      error: () => alert("Erro ao atualizar usuário")
+  ngAfterInit(){
+     this.userService.getUser$().subscribe(u => {
+        this.user = u;
+        this.loading = false;
     });
   }
 
+  onEdit() {
+    if(this.user){
+      this.userService.editUser(this.user.id, this.user).subscribe({
+        next: () => alert("Usuário atualizado com sucesso"),
+        error: () => alert("Erro ao atualizar usuário")
+      });
+    }
+  }
+
   onDelete() {
-    if(confirm("Tem certeza que deseja excluir a conta?")){
+    if(confirm("Tem certeza que deseja excluir a conta?") && this.user){
       this.userService.deleteUser(this.user.id).subscribe({
         next: () => alert("Usuário excluído com sucesso"),
         error: () => alert("Erro ao excluir usuário")
